+++ src/features/implementationProvider.ts
@@ -6,7 +6,7 @@
 import AbstractSupport from './abstractProvider';
 import { FindImplementationsRequest } from '../omnisharp/protocol';
 import * as serverUtils from '../omnisharp/utils';
-import { createRequest, toLocation } from '../omnisharp/typeConvertion';
+import { createRequest, toLocation } from '../omnisharp/typeConversion';
 import { TextDocument, Position, CancellationToken, ImplementationProvider, ProviderResult, Definition } from 'vscode';
 
 export default class CSharpImplementationProvider extends AbstractSupport implements ImplementationProvider {
