+++ src/features/referenceProvider.ts
@@ -6,7 +6,7 @@
 import AbstractSupport from './abstractProvider';
 import * as protocol from '../omnisharp/protocol';
 import * as serverUtils from '../omnisharp/utils';
-import {createRequest, toLocation} from '../omnisharp/typeConvertion';
+import {createRequest, toLocation} from '../omnisharp/typeConversion';
 import {ReferenceProvider, Location, TextDocument, CancellationToken, Position} from 'vscode';
 
 export default class OmnisharpReferenceProvider extends AbstractSupport implements ReferenceProvider {
