+++ src/features/signatureHelpProvider.ts
@@ -5,7 +5,7 @@
 
 import AbstractSupport from './abstractProvider';
 import * as serverUtils from '../omnisharp/utils';
-import { createRequest } from '../omnisharp/typeConvertion';
+import { createRequest } from '../omnisharp/typeConversion';
 import { SignatureHelpProvider, SignatureHelp, SignatureInformation, ParameterInformation, CancellationToken, TextDocument, Position } from 'vscode';
 import { MarkdownString } from 'vscode';
 import { SignatureHelpParameter } from '../omnisharp/protocol';
