+++ src/omnisharp/typeConversion.ts
